requirements.txt
pytorch>=0.4.1
numpy>=1.15.4